other half is for future SDSS paper
