+++ case_studies/sdss_galaxies/README.md
@@ -1,5 +1,5 @@
 sdss_galaxies
 ========
-Contains scripts to re-run the SDSS galaxies experiment with the autoencoder.
+Contains scripts to re-run the SDSS galaxies experiment.
 
-- `plots.py` contains code for reproducing figures for the paper.
+- `figures.py` contains code for reproducing figures for the paper.
