+++ case_studies/sdss_galaxies/config/config.yaml
@@ -3,8 +3,6 @@ defaults:
     - training: sdss_autoencoder
     - override hydra/job_logging: stdout
 
-# completely disable hydra logging
-# https://github.com/facebookresearch/hydra/issues/910
 hydra:
     output_subdir: null
     run:
@@ -20,14 +18,6 @@ paths:
     output: ${paths.project}/output
 
 datasets:
-    sdss_blended_galaxies:
-        _target_: bliss.datasets.sdss_blended_galaxies.SdssBlendedGalaxies
-        detection_encoder: ${models.detection_encoder}
-        binary_encoder: ${models.binary}
-        location_ckpt: ${paths.project}/models/sdss_detection_encoder.pt
-        binary_ckpt: ${paths.project}/models/sdss_binary.pt
-        sdss_dir: ${paths.sdss}
-        prerender_device: "cuda"
     sdss_galaxies:
         _target_: bliss.datasets.galsim_galaxies.SingleGalsimGalaxies
         prior:
@@ -60,28 +50,6 @@ datasets:
         n_batches: 10
         fix_validation_set: True
         valid_n_batches: 200
-    galsim_blends:
-        _target_: bliss.datasets.galsim_galaxies.GalsimBlends
-        prior:
-            _target_: bliss.models.galsim_decoder.UniformGalsimPrior
-            single_galaxy_prior: ${datasets.sdss_galaxies.prior}
-            max_n_sources: 8
-            mean_sources: 3
-            max_shift: 0.5
-            galaxy_prob: 0.7
-        decoder:
-            _target_: bliss.models.galsim_decoder.FullCatalogDecoder
-            single_galaxy_decoder: ${datasets.sdss_galaxies.decoder}
-            slen: 40
-            bp: 24
-        background: ${datasets.sdss_galaxies.background}
-        tile_slen: 4
-        max_sources_per_tile: 1
-        num_workers: 10
-        batch_size: 32
-        n_batches: 10
-        fix_validation_set: True
-        valid_n_batches: 50
     simulated:
         _target_: bliss.datasets.simulated.SimulatedDataset
         decoder: ${models.decoder}
@@ -101,22 +69,6 @@ datasets:
         generate_device: "cuda:0"
         fix_validation_set: true
         valid_n_batches: 256
-
-    toy_gaussian:
-        _target_: bliss.datasets.galsim_galaxies.ToyGaussian
-        num_workers: 0
-        n_batches: 10
-        batch_size: 64
-        slen: 53
-        n_bands: 1
-        background: 865.
-        psf_fwhm: 0.8
-        pixel_scale: 0.4
-        min_flux: 300
-        max_flux: 10000
-        min_hlr: 0.8
-        max_hlr: 3.0
-        max_e: 0.6
 
 models:
     binary:
@@ -190,11 +142,6 @@ models:
             latents_file: ${paths.project}/models/latents_simulated_sdss_galaxies.pt
             galaxy_dataset: ${datasets.sdss_galaxies}
 
-generate:
-    dataset:
-    file:
-    n_plots:
-
 training:
     n_epochs: 121
     experiment: default
@@ -222,40 +169,17 @@ training:
     seed: 42
 
 plots:
+    figs:
+        - "sdss_detection"
+        - "sdss_recon"
     location_checkpoint: ${paths.project}/models/sdss_detection_encoder.pt
     galaxy_checkpoint: ${paths.project}/models/sdss_galaxy_encoder.pt
     binary_checkpoint: ${paths.project}/models/sdss_binary.pt
-    figs:
-        - 1
-        - 2
-        - 3
     cachedir: ${paths.project}/output/sdss_figures_cache
     figdir: ${paths.project}/figures
     image_format: "png"
     overwrite: false
     psf_params_file: ${paths.data}/sdss/94/1/12/psField-000094-1-0012.fits
-    simulated_sdss_individual_galaxies: ${paths.project}/models/simulated_sdss_individual_galaxies.pt
-    simulated_blended_galaxies: ${paths.project}/models/simulated_blended_galaxies.pt
-    galsim_blended_galaxies:
-        _target_: bliss.datasets.galsim_galaxies.GalsimBlends
-        prior:
-            _target_: bliss.models.galsim_decoder.UniformGalsimPrior
-            single_galaxy_prior: ${datasets.sdss_galaxies.prior}
-            max_n_sources: 8
-            mean_sources: 3
-            max_shift: 0.375
-            galaxy_prob: 1.0
-        decoder:
-            _target_: bliss.models.galsim_decoder.FullCatalogDecoder
-            single_galaxy_decoder: ${datasets.sdss_galaxies.decoder}
-            slen: 40
-            bp: 24
-        background: ${datasets.sdss_galaxies.background}
-        tile_slen: 4
-        max_sources_per_tile: 1
-        num_workers: 20
-        batch_size: 512
-        n_batches: 20
     sdss_pixel_scale: 0.396
     device: "cuda:0"
     coadd_file: ${paths.sdss}/coadd_catalog_94_1_12.fits
