--- case_studies/sdss_galaxies/config/generate/sdss_individual_galaxies.yaml
@@ -1,14 +1,0 @@
-# @package _global_
-generate:
-  dataset: ${datasets.sdss_galaxies}
-  file: data/simulated_sdss_individual_galaxies
-  common:
-    - background
-    - slen
-
-datasets:
-  sdss_galaxies:
-    batch_size: 512
-    n_batches: 20
-    num_workers: 20
-    flux_sample: uniform
