--- case_studies/sdss_galaxies/models/latents_simulated_sdss_galaxies.pt
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:c55d22a6c813d96dd4bb639fe4f992fbc35127f4dddc606fcaf4d42a7a107198
-size 1311603
