--- case_studies/sdss_galaxies/models/sdss_autoencoder.pt
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:02b8cde82da7346a719ba7c28494967282f02250f9d4d40a52d154ab56567caf
-size 442019
