--- case_studies/sdss_galaxies/models/sdss_autoencoder.pt.log.json
@@ -1,1 +1,0 @@
-{"timestamp": "2022-11-28 03:15:46.143927", "train_time_sec": 21024.900178217, "monitor": "val/loss", "best_model_score": 3460202.0, "best_model_path": "/home/imendoza/bliss/case_studies/sdss_galaxies/output/sdss_autoencoder/version_3/checkpoints/epoch=999-val_loss=3460202.000.ckpt", "current_score": 3460202.0, "dirpath": "/home/imendoza/bliss/case_studies/sdss_galaxies/output/sdss_autoencoder/version_3/checkpoints", "kth_best_model_path": "/home/imendoza/bliss/case_studies/sdss_galaxies/output/sdss_autoencoder/version_3/checkpoints/epoch=999-val_loss=3460202.000.ckpt", "kth_value": 3460202.0, "last_model_path": ""}
