--- case_studies/sdss_galaxies/models/sdss_binary.pt
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:058720548736e752994d86de074d1277ed2a500d0608b012845f33f225f675fe
-size 2383439
