--- case_studies/sdss_galaxies/models/sdss_binary.pt.log.json
@@ -1,1 +1,0 @@
-{"timestamp": "2022-11-28 11:44:15.590623", "train_time_sec": 10971.250195409, "monitor": "val/loss", "best_model_score": 22.55724334716797, "best_model_path": "/home/imendoza/bliss/case_studies/sdss_galaxies/output/sdss_binary/version_2/checkpoints/epoch=489-val_loss=22.557.ckpt", "current_score": 22.55724334716797, "dirpath": "/home/imendoza/bliss/case_studies/sdss_galaxies/output/sdss_binary/version_2/checkpoints", "kth_best_model_path": "/home/imendoza/bliss/case_studies/sdss_galaxies/output/sdss_binary/version_2/checkpoints/epoch=489-val_loss=22.557.ckpt", "kth_value": 22.55724334716797, "last_model_path": ""}
