--- case_studies/sdss_galaxies/models/sdss_detection_encoder.pt
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:222909edd9fa17741d060921ceb07e2823bb2be3efca6e5e835b1bc902d00863
-size 3175731
