--- case_studies/sdss_galaxies/models/sdss_detection_encoder.pt.log.json
@@ -1,1 +1,0 @@
-{"timestamp": "2022-11-29 03:12:18.943808", "train_time_sec": 47902.496847502, "monitor": "val/loss", "best_model_score": -0.026575539261102676, "best_model_path": "/home/imendoza/bliss/case_studies/sdss_galaxies/output/sdss_detection_encoder/version_3/checkpoints/epoch=1399-val_loss=-0.027.ckpt", "current_score": -0.026575539261102676, "dirpath": "/home/imendoza/bliss/case_studies/sdss_galaxies/output/sdss_detection_encoder/version_3/checkpoints", "kth_best_model_path": "/home/imendoza/bliss/case_studies/sdss_galaxies/output/sdss_detection_encoder/version_3/checkpoints/epoch=1399-val_loss=-0.027.ckpt", "kth_value": -0.026575539261102676, "last_model_path": ""}
