--- case_studies/sdss_galaxies/models/sdss_galaxy_encoder.pt
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:597d09322a4c8e4246e21bf7caaf7abf187ead5982d819cd1e546979179edcfc
-size 371999
