--- case_studies/sdss_galaxies/models/sdss_galaxy_encoder.pt.log.json
@@ -1,1 +1,0 @@
-{"timestamp": "2022-11-28 13:53:48.872838", "train_time_sec": 7764.020186251, "monitor": "val/loss", "best_model_score": 1194144.75, "best_model_path": "/home/imendoza/bliss/case_studies/sdss_galaxies/output/sdss_galaxy_encoder/version_2/checkpoints/epoch=3949-val_loss=1194144.750.ckpt", "current_score": 1194144.75, "dirpath": "/home/imendoza/bliss/case_studies/sdss_galaxies/output/sdss_galaxy_encoder/version_2/checkpoints", "kth_best_model_path": "/home/imendoza/bliss/case_studies/sdss_galaxies/output/sdss_galaxy_encoder/version_2/checkpoints/epoch=3949-val_loss=1194144.750.ckpt", "kth_value": 1194144.75, "last_model_path": ""}
