--- case_studies/sdss_galaxies/models/simulated_sdss_individual_galaxies.pt
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:86b5c63539c9f7ffc33734eddf8a17fa5c53179ab13729a4a8b8d0e2c713634a
-size 236236888
