+++ case_studies/strong_lensing/plots/main.py
@@ -13,10 +13,10 @@ from bliss.catalog import PhotoFullCatalog
 from bliss.encoder import Encoder
 from bliss.inference import SDSSFrame, SimulatedFrame
 from bliss.models.decoder import ImageDecoder
-from case_studies.sdss_galaxies.plots.autoencoder import AEReconstructionFigures
-from case_studies.sdss_galaxies.plots.sdss_detection_metrics import DetectionClassificationFigures
+from case_studies.sdss_galaxies.plots.ae_recon import AEReconstructionFigures
+from case_studies.sdss_galaxies.plots.galsim_blends import BlendSimFigures
+from case_studies.sdss_galaxies.plots.sdss_detection import DetectionClassificationFigures
 from case_studies.sdss_galaxies.plots.sdss_reconstruction import SDSSReconstructionFigures
-from case_studies.sdss_galaxies.plots.sim_blend_metrics import BlendSimFigures
 
 
 def load_models(cfg, device):
